replicate: use bearer token instead of did (#111)
The previous DID stuff in the replicate crates were unecessary and too
application specific. Its more useful to use a general bearer token, and
leave how to validate that token and what its contents are up to the
application.

A did-signed authentication attestation could just be a base64 signed
message, and use that as the bearer token.