physnet: connect to instance (#116)
Implements the last bits of functionality in the title screen. Nothing
is faked anymore. Both connections (Manager, Instance) are actually
done. 🥳


https://github.com/NexusSocial/nexus-vr/assets/6969415/a6f6518f-71a4-465d-bc49-96c4392891e9




Next step will be to implement the background task that syncs from
client to server for the instance. In other words, I think the entire
demo is done now, and we can just focus on implementing the state sync
in the networking crate.